Working front.

+++ Bets/BetsData/Entities/Account.cs
@@ -8,6 +8,6 @@ namespace BetsData.Entities
     {
         public string UserId { get; set; }
 
-        public long Balance { get; set; }
+        public double Balance { get; set; }
     }
 }
